test(cubesql): Extract user change tests to separate module
